Moved implicits package to coulomb.cats
